+++ src/components/common/modals/tertiaryModal/CloseButton.tsx
@@ -1,7 +1,7 @@
 import CloseButton from '@components/common/buttons/CloseButton';
 import { SetStateAction } from 'react';
 import styled from 'styled-components';
-import { MODAL_CLOSE_TRANSITION_TIME } from '@lib/constants';
+import { MODAL_CLOSE_TRANSITION_TIME } from '@lib/constants/MODAL';
 
 interface CloseButtonProps {
   shouldCloseOnOverlayClick: boolean;
