+++ src/components/mypage/OverView.tsx
@@ -3,7 +3,7 @@ import renderRoleIllust from '@lib/utils/render/renderRoleIllust';
 import { IGetUserResData } from '@lib/api/user/user.type';
 import OverViewContent from './OverViewContent';
 import { useQueryClient } from 'react-query';
-import { KID } from '@lib/constants/queryKeys';
+import { KID } from '@lib/constants/QUERY_KEY';
 import getPercentValue from '@lib/utils/get/getPercentValue';
 import { IKidListDTO } from '@lib/api/family/family.type';
