+++ src/lib/constants/AWS_S3_URL.ts
@@ -1,0 +1,1 @@
+export const AWS_S3_URL = `${process.env.REACT_APP_AWS_S3_URL}`;
