+++ src/lib/constants/BASE_URL.ts
@@ -1,0 +1,2 @@
+// export const BASE_URL = 'https://api.bankidz.com'; // 배포 환경
+export const BASE_URL = 'https://bankids.click'; // 테스트 환경
