+++ src/lib/constants/MODAL.ts
@@ -1,0 +1,3 @@
+export const MODAL_CLOSE_TRANSITION_TIME = 200; // milliseconds
+export const MODAL_SLIDE_FROM_POSITION = '-10%';
+export const MODAL_SLIDE_TO_POSITION = '-50%';
